% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework2.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Evaluate a multivariate normal (log-)density function at a specified point
using Cholesky or eigen decomposition}
\usage{
dmvnorm(x, mean, cov, method = c("chol", "eigen"), log = TRUE)
}
\arguments{
\item{x}{A vector, the point at which to evaluate the density.}

\item{mean}{The mean vector for the multivariate distribution.}

\item{cov}{The covariance matrix for the multivariate distribution.}

\item{method}{The method to use for matrix decomposition, either "chol" for
Cholesky decomposition or "eigen" for eigen decomposition. Defaults to "chol".}

\item{log}{A boolean, TRUE to return log-density and FALSE to
return the density. Defaults to TRUE.}
}
\value{
A single numeric value, the (log-)density at the specified point.
}
\description{
Evaluate a multivariate normal (log-)density function at a specified point
using Cholesky or eigen decomposition
}
\examples{
density <- dmvnorm(x = rep(0, 4), mean = rep(0, 4), cov = diag(1, nrow = 4))
}
