% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{mult_ABx}
\alias{mult_ABx}
\title{Multiply two square matrices and a vector}
\usage{
mult_ABx(A, B, x, slow = FALSE)
}
\arguments{
\item{A}{A square matrix.}

\item{B}{Another square matrix.}

\item{x}{A vector.}

\item{slow}{A boolean indicating whether to multiply using the slower method,
\eqn{(AB)x}, or the faster method, \eqn{A(Bx)} (default).}
}
\value{
A numeric vector or column/row vector (\eqn{1\times n} or \eqn{n \times 1}
matrix).
}
\description{
Multiply two square matrices and a vector
}
\examples{
mult_ABx(matrix(rnorm(25), nrow = 5), matrix(runif(25, 0, 10), nrow = 5), rpois(5, 5))
}
