% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{mult_Ainv_c}
\alias{mult_Ainv_c}
\title{Multiply the inverse of a square matrix by a vector}
\usage{
mult_Ainv_c(A, c, direct = TRUE)
}
\arguments{
\item{A}{A square matrix.}

\item{c}{A vector.}

\item{direct}{A boolean indicating whether to solve the system directly using
\eqn{Ax = c} (default), or compute \eqn{A^{-1}c} by first inverting A.}
}
\value{
A numeric vector or column/row vector (\eqn{1\times n} or \eqn{n \times 1}
matrix).
}
\description{
Multiply the inverse of a square matrix by a vector
}
\examples{
mult_Ainv_c(matrix(rnorm(25), nrow = 5), rpois(5, 5))
}
