% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework2.R
\name{my_lm}
\alias{my_lm}
\title{Use Cholesky decomposition, QR decomposition, or Singular Value Decomposition
to get OLS estimates}
\usage{
my_lm(formula, data, method = c("chol", "qr", "svd"), se = FALSE)
}
\arguments{
\item{formula}{A formula object of the form y ~ x1 + x2 + ... that specifies
the model to be fit.}

\item{data}{The dataframe containing the data for the OLS fit, including a
column of 1s if an intercept is desired.}

\item{method}{Methods available are "chol" for Cholesky decomposition, "qr"
for QR decomposition, and "svd" for Singular Value Decomposition.}

\item{se}{A boolean, TRUE to return standard errors for coefficients.
Defaults to FALSE.}
}
\value{
A named vector containing coefficient names and their corresponding
estimates. If se = TRUE, returns a list with elements "coefficients" and
"standard_errors".
}
\description{
Use Cholesky decomposition, QR decomposition, or Singular Value Decomposition
to get OLS estimates
}
