% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework2.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulate realizations from a multivariate normal distribution using
Cholesky decomposition}
\usage{
rmvnorm(mean, cov, N)
}
\arguments{
\item{mean}{The mean vector for the multivariate distribution.}

\item{cov}{The covariance matrix for the multivariate distribution.}

\item{N}{The number of realizations to be returned.}
}
\value{
A matrix with columns of realizations of the multivariate normal.
}
\description{
Simulate realizations from a multivariate normal distribution using
Cholesky decomposition
}
\examples{
sample <- rmvnorm(mean = rep(0, 4), cov = diag(1, nrow = 4), N = 100)
}
