% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5.R
\name{wright_fisher_path}
\alias{wright_fisher_path}
\title{Wright-Fisher Model Path Simulation}
\usage{
wright_fisher_path(n_steps, x0, N = 100, u = 0.35, v = 0.5)
}
\arguments{
\item{n_steps}{Integer. Number of time steps to simulate.}

\item{x0}{Integer. Initial number of copies of the allele (must be between 0 and N).}

\item{N}{Integer. Population size (number of alleles). Default is 100.}

\item{u}{Numeric. Mutation rate from allele type 1 to type 0 (between 0 and 1).
Default is 0.35.}

\item{v}{Numeric. Mutation rate from allele type 0 to type 1 (between 0 and 1).
Default is 0.5.}
}
\value{
Numeric vector of length n_steps + 1 containing the number of copies of
the allele at each time step, starting with x0.
}
\description{
Simulates a path of allele frequencies under the Wright-Fisher model
with mutation.
}
